Don't apply /v1 prefix to .well-known or related files.
