+++ routes/idp.js
@@ -39,30 +39,6 @@ module.exports = function (log, crypto, error, isA, serverPublicKey, bridge) {
           file: './provision.html'
         }
       }
-    },
-    {
-      method: 'POST',
-      path: '/get_random_bytes',
-      config: {
-        handler: function getRandomBytes(request) {
-          request.reply({ data: crypto.randomBytes(32).toString('hex') })
-        }
-      }
-    },
-    {
-      method: 'GET',
-      path: '/verify_email',
-      config: {
-        handler: function (request, next) {
-          return request.reply.redirect(bridge.url + request.raw.req.url)
-        },
-        validate: {
-          query: {
-            code: isA.String().regex(HEX_STRING).required(),
-            uid: isA.String().max(64).required()
-          }
-        }
-      }
     }
   ]
